backend/requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.1.0
transformers==4.35.2
huggingface-hub==0.16.4
beautifulsoup4==4.12.2
requests==2.31.0 